additional tests, contract reorganization
